requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
pydantic-settings==2.1.0
duckduckgo-search==4.1.1
python-dotenv==1.0.0
httpx==0.26.0
